+++ maverics/src/main/kotlin/com/example/model/OrderHistory.kt
@@ -1,9 +1,11 @@
 package com.example.model
 
+import java.math.BigInteger
+
 data class BuyOrderHistory(
     val orderId: Int,
-    val price: Long,
-    val quantity: Long,
+    val price: BigInteger,
+    val quantity: BigInteger,
     val type: String,
     val status: String = "unfilled",
     val filled: MutableList<Transaction>
@@ -11,8 +13,8 @@ data class BuyOrderHistory(
 
 data class SellOrderHistory(
     val orderId: Int,
-    val price: Long,
-    val quantity: Long,
+    val price: BigInteger,
+    val quantity: BigInteger,
     val type: String,
     val esopType: String? = null,
     val status: String = "unfilled",
